+++ frontend/src/assets/mock.js
@@ -1,0 +1,81 @@
+export const produtos = [
+  {
+    imageSrc: "/produtos/topper.webp",
+    title: "Toppers Fauna Brasileira",
+    price: "R$ 50,00",
+    productLink: "https://example.com/topper"
+  },
+  {
+    imageSrc: "/produtos/topper.webp",
+    title: "Toppers Fauna Brasileira",
+    price: "R$ 50,00",
+    productLink: "https://example.com/topper"
+  },
+  {
+    imageSrc: "/produtos/topper.webp",
+    title: "Toppers Fauna Brasileira",
+    price: "R$ 50,00",
+    productLink: "https://example.com/topper"
+  },
+  {
+    imageSrc: "/produtos/topper.webp",
+    title: "Toppers Fauna Brasileira",
+    price: "R$ 50,00",
+    productLink: "https://example.com/topper"
+  },
+  {
+    imageSrc: "/produtos/topper.webp",
+    title: "Toppers Fauna Brasileira",
+    price: "R$ 50,00",
+    productLink: "https://example.com/topper"
+  },
+  {
+    imageSrc: "/produtos/topper.webp",
+    title: "Toppers Fauna Brasileira",
+    price: "R$ 50,00",
+    productLink: "https://example.com/topper"
+  },
+  {
+    imageSrc: "/produtos/topper.webp",
+    title: "Toppers Fauna Brasileira",
+    price: "R$ 50,00",
+    productLink: "https://example.com/topper"
+  },
+];
+
+export const depoimentos = [
+  {
+    testimony: "Minha encomenda chegou até antes da hora, impecavelmente perfeita, tudo lindo e muito caprichoso, até o cuidado com a embalagem foi um diferencial! A Léa é super atenciosa e foi capaz de “traduzir” em arte de papel o meu desejo! Parabéns!",
+    author: "Lisandra",
+  },
+  {
+    testimony: "O atendimento foi impecável desde o primeiro momento. A encomenda foi perfeita. De uma delicadeza e de uma beleza. Feito com muito carinho. E ainda enviou brinde. Minha filha e eu adoramos. Muito obrigada!",
+    author: "Claudia",
+  },
+  {
+    testimony: "Os produtos são de excelente qualidade e o atendimento totalmente personalizado. Todas as alterações solicitadas foram atendidas. Fiquei muito satisfeita, mais uma vez!",
+    author: "Simone",
+  },
+];
+
+export const images = [
+    {
+        src: "images/hero-final.png"
+    },
+    {
+        src: "images/hero-final.png"
+    },
+    {
+        src: "images/hero-final.png"
+    },
+]
+
+
+
+
+
+
+
+
+
+