+++ frontend/CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added responsive Hero
+- Created Testimonies
+- Added React-Icons
 - Created ProductCard and ProductCarousel, using the Swiper lib
 - Added custom colors and fonts to Tailwind in index.css
 - Added page title and favicon in "index.html"